+++ Src/Application/LoadingState.h
@@ -90,7 +90,7 @@ namespace Application
 		la última llamada (o desde la áctivación del estado, en caso
 		de ser la primera vez...).
 		*/
-		virtual void tick(unsigned int msecs);
+		virtual void tick(float msecs);
 
 		/**
 		Cambia el estado que se va a cargar tras el estado de loading.
