+++ Src/Logic/Entity/Components/Camera.h
@@ -94,7 +94,7 @@ namespace Logic
 
 		@param msecs Milisegundos transcurridos desde el último tick.
 		*/
-		virtual void tick(unsigned int msecs);
+		virtual void tick(float msecs);
 
 	protected:
