+++ Src/Logic/Entity/Components/LeverInteractuable.h
@@ -17,7 +17,7 @@ namespace Logic
 
 		bool spawn(const std::string& name, CEntity* entity, CMap *map, const Map::CEntity *entityInfo) override;
 
-		void tick(unsigned int msecs) override;
+		void tick(float msecs) override;
 
 		virtual bool activate() override;
