+++ Src/Logic/Entity/Components/Patrol.h
@@ -29,7 +29,7 @@ namespace Logic
 		
 		void process(const TMessage &message) override;
 
-		void tick(unsigned int msecs) override;
+		void tick(float msecs) override;
 
 	protected:
