+++ Src/Logic/Maps/Map.cpp
@@ -290,7 +290,7 @@ namespace Logic {
 
 	//---------------------------------------------------------
 
-	void CMap::tick(unsigned int msecs) 
+	void CMap::tick(float msecs)
 	{
 		TGameObjectMap::const_iterator it;
