+++ Src/Logic/Maps/Map.h
@@ -113,7 +113,7 @@ namespace Logic
 
 		@param msecs Milisegundos transcurridos desde el último tick.
 		*/
-		void tick(unsigned int msecs);
+		void tick(float msecs);
 
 		/**
 		Añade una nueva entidad al mapa. Si la entidad ya estaba incluida
